Automate BeagleBone Yocto image build process
Changes:
- Updated build process to use the current working directory instead of a hardcoded path
- Added check for existing Yocto directory with overwrite/exit prompt
- Integrated custom recipe into image build
- Implemented timestamped log file generation for each run
- Renamed script for clarity and consitency

Test Done:
- Verified Yocto directory is created if not present
- Verified prompt appears when directory already exists
- Confirmed image builds successfully with `bitbake core-image-cjson` without separate recipe build step
- Confirmed log file is created with correct timestamp for each run

Known Issues:
- Terminal output mirrors the log file content instead of showing the usual execution output

+++ Yocto_install/yocto-custom-build.sh
@@ -42,7 +42,7 @@ sudo apt install -y \
 
 echo "============= Generate locale  ================"
 sudo locale-gen en_US.UTF-8
-do update-locale LANG=en_US.UTF-8 LC_ALL=en_US.UTF-8
+sudo update-locale LANG=en_US.UTF-8 LC_ALL=en_US.UTF-8
 
 export LANG=en_US.UTF-8
 export LC_ALL=en_US.UTF-8
@@ -87,13 +87,13 @@ fi
 echo "=============== Creating cloning recipes from git  ==============="
 cd $BUILD_DIR/meta-jsonapp/
 
-git clone https://github.com/rathod-om/Practice_repo
+if [ ! -d recipes-jsonapp ]; then
+	git clone https://github.com/rathod-om/Practice_repo
 
-
-cp -r $BUILD_DIR/meta-jsonapp/Practice_repo/ravi/recipes-jsonapp/ .
-
-rm -rf $BUILD_DIR/meta-jsonapp/Practice_repo  
-
+	cp -r $BUILD_DIR/meta-jsonapp/Practice_repo/ravi/recipes-jsonapp/ .
+	
+	rm -rf $BUILD_DIR/meta-jsonapp/Practice_repo  
+fi
 
 echo "=============== bitbaking the core-image-cjson code  ==============="
